Ajustements jeux
- Stats mauvaises
- messAd inexistant dans certains cas
- Mise affichée dans OT!tortues CS

+++ Jeux/CrossServeur/BRCross.py
@@ -97,4 +97,5 @@ async def trivialBattleRoyaleCross(ctx,bot,inGame,gamesTrivial):
         await ctx.send(embed=embedAssert(er))
     except:
         await game.error(ctx,bot,message,inGame,gamesTrivial)
-    await messAd.delete()
+    if "messAd" in locals():
+        await messAd.delete()
