+++ Jeux/Trivial/BattleRoyale.py
@@ -148,4 +148,5 @@ async def trivialBattleRoyale(ctx,bot,inGame,gamesTrivial):
         await ctx.send(embed=embedAssert(er))
     except:
         await game.error(ctx,bot,message,inGame,gamesTrivial)
-    await messAd.delete()
+    if "messAd" in locals():
+        await messAd.delete()
